pyproject.toml
[project]
name = "sabobot"
version = "0.1.0"
description = "Sabobot - ElevenLabs Conversational AI Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]
